requirements.txt
discord.py==1.7.3
icalevents==0.1.25
